Python: fixed error reporting on initialization of applications.
PyErr_Print() writes traceback to "sys.stderr", which is a file object that
can buffer the output.  If the process exits immediately, the buffer can be
destroyed before flushing to the log.  As a result, the user doesn't see
the traceback.

Now Py_Finalize() is also called in case of any errors during initialization.
It finalizes the interpreter and flushes all data.