Add capability/privilege checking
Current behaviour ignores the EPERM error when
setuid/setgid to the unprivileged process.
There are two cases:

  - Unit unprivileged processes (discovery, controller and router)
  - User processes

Currently, Unit ignores the error in both cases. This is difficult
to keep as very soon we are going to support executing isolated
applications. If Unit silently ignores the user specified in the
application spec, then the user process can run with more capabilities
than needed, leading to critical vulnerabilities (eg.: access granted to files/dirs
that otherwise could not be possible).

This commit is partial, because some discussion is required
about default user and group.

Signed-off-by: Tiago Natel de Moura <tiago.moura@nginx.com>