Tests: fixing racing condition in ASGI threads test.
ASGI threads read all the requests from the queue before start processing it.
This why test need to wait a little to let the ASGI thread start request
processing and block.  In virtual environment any thread or process may
be delayed and only method to avoid racing is a reasonable sleep increase.