Python: using default event_loop for main thread for ASGI.
Unit's ASGI implementation creates a new event loop to run an application for
each thread since 542b5b8c0647.  This may cause unexpected exceptions or
strange bugs if asyncio synchronisation primitives are initialised before the
application starts (e.g. globally).

Although the approach with a new event loop for the main thread is consistent
and helps to prepare the application to run in multiple threads, it can be a
source of pain for people who just want to run single-threaded ASGI
applications in Unit.

This is related to #560 issue on GitHub.